adjust setup script a little more

+++ setup.sh
@@ -6,16 +6,18 @@ git submodule update --init
 cd msgpack-c
 git checkout cpp_master
 
-cd /tmp/
-git clone https://github.com/mfontanini/libtins
-cd libtins
-mkdir build
-cd build
-cmake ../ -DLIBTINS_ENABLE_CXX11=1
-make -j$(nproc)
-make install
+if [ ! -f /usr/local/lib/libtins.so ]; then
+  cd /tmp/
+  git clone https://github.com/mfontanini/libtins
+  cd libtins
+  mkdir build
+  cd build
+  cmake ../ -DLIBTINS_ENABLE_CXX11=1
+  make -j$(nproc)
+  make install
 
-cd /tmp/
-rm -rf libtins
+  cd /tmp/
+  rm -rf libtins
+fi
 
 ldconfig
